added a new language

+++ src/views/AdminSigninView.vue
@@ -1,30 +1,29 @@
 <template>
   <div class="auth-container">
     <div class="form-l-wrapper">
-      <h1>{{$t('auth.h1-1')}}</h1>
+      <h1>Create an account</h1>
       <form @submit.prevent="create" class="l-form">
         <input type="text" class="input-l" placeholder="Full Name" v-model="username" />
         <input type="email" class="input-l" placeholder="Email Address" v-model="email" />
         <select class="input-l" v-model="selectedCountry">
-          <option disabled value="">{{$t('auth.op')}}</option>
+          <option disabled value="">Select a country</option>
           <option v-for="country in countriesData" :key="country.code" :value="country.code">
             {{ country.name }}
           </option>
         </select>
         <input type="password" class="input-l" placeholder="Password" v-model="password" />
-        <input type="password" class="input-l" placeholder="Confirm password" v-model="confirmPassword" />
+        <input
+          type="password"
+          class="input-l"
+          placeholder="Confirm password"
+          v-model="confirmPassword"
+        />
         <p>{{ errMsg }}</p>
-        <button class="btn-f" type="submit">{{$t('auth.btn-2')}}</button>
+        <button class="btn-f" type="submit">Sign up</button>
       </form>
       <span>or</span>
       <div class="l-alternatives">
-        <button class="alt-btn" @click="login">
-          {{$t('auth.btn-1')}}
-        </button>
-         <!-- <div class="auth-google-contain" @click="loginInWithGoogle">
-          <googleIcon class="auth-google" />
-          <span> sign in with google</span>
-        </div> -->
+        <button class="alt-btn" @click="login">Log in</button>
       </div>
     </div>
   </div>
@@ -35,15 +34,15 @@ import { ref } from 'vue'
 import axios from 'axios'
 import { useRouter } from 'vue-router'
 import { useAuthStore } from '../stores/auth'
+import { useToast } from 'vue-toastification'
 import countriesData from '../components/countries.json'
-// import googleIcon from '../icons/googleIcon.vue'
+
 const SERVER_HOST = import.meta.env.VITE_SERVER_HOST
-
-
 const selectedCountry = ref('')
 const router = useRouter()
 const username = ref('')
 const password = ref('')
+const toast = useToast()
 const errMsg = ref('')
 const email = ref('')
 const confirmPassword = ref('')
@@ -58,36 +57,40 @@ const reset = () => {
 }
 
 const create = async () => {
+  if (password.value !== confirmPassword.value) {
+    toast.error('Passwords do not match')
+    return
+  }
+
   if (username.value !== '' && password.value !== '') {
     try {
       const response = await axios.post(`${SERVER_HOST}/auth/register-admin`, {
         username: username.value,
         email: email.value,
         password: password.value,
+        country: selectedCountry.value,
+        confirmPassword: confirmPassword.value,
         selectedCountry: selectedCountry.value
       })
-      console.log(response.data) 
       const token = response.data.token
       const id = response.data._id
       authStore.toggleToken(JSON.stringify(token))
-      
       localStorage.setItem('username', username)
       localStorage.setItem('token', JSON.stringify(token))
       localStorage.setItem('id', id)
-
+      localStorage.setItem('admin', true)
       router.push({ name: 'Panel' })
+      toast.success('Welcome Admin!')
     } catch (error) {
-      console.error(error)
+      toast.error(error.response.data.message)
     }
   } else {
-    errMsg.value = 'Write something'
+    toast.error('Please enter all the required fields')
+
     reset()
   }
 }
 
-// const loginInWithGoogle = () => {
-  
-// }
 const login = () => {
   router.push({ name: 'AdminLogin' })
 }
